If we load a dialplan with a hint referring to a device state provider that is not yet registered, the device state for the device in the hint will not be initialized. For this solution, that means we can't set up a subscription.
Therefore I'm adding a reinitialization function run when a device state provider registers.
If we already have hints using this provider, we recheck their state. This way we get a first 
state and can set up services needed for the device state provider.

Actually proved that it worked to. Amazing.