main/features: Use POLLPRI when waiting for events on parked channels.
This change resolves a regression in the 1.6.2 when converting from
select to poll.  The DAHDI timers use POLLPRI to indicate that the timer
fired, but features was not waiting for that flag.  The result was no
audio for MOH when a call was parked and res_timing_dahdi was in use.

This patch is slightly modified from the one on the mantis issue.  It does
not set an exception on the channel if the POLLPRI flag is set.

(closes issue #18262)
Reported by: francesco_r
Patches:
      patch_park_moh-trunk-2.txt uploaded by cjacobsen (license 1029)
      Tested by: francesco_r, rfrantik, one47