New branch for DTMF work
Asterisk currently doesn't send out RTP DTMF with the same duration as the incoming.
There was a lot of bugs in this code, but the most important one was the lack
of duration being sent from the incoming call leg to the outbound.

This code will add a new control frame for the continuation messages and make
sure that our DTMF end packets will not be far away from the incoming. There are
some adjustments done since we clock outbound DTMF on inbound RTP packets.