Add AELSub, which provides a stable entry point into AEL subroutines.
This commit needs some explanation, given that we're adding a new application
into an existing release branch.  This is generally a violation of our release
policy, except in very limited circumstances, and I believe this is one of
those circumstances.

The problem that this solves is one of the sanity of using multiple dialplan
languages to define a dialplan.  In the case of the reporter, he or she is
using AEL is define subroutines, while using Realtime extensions to invoke
those subroutines.  While you can do this, it's based upon the reality of AEL
using actual dialplan extensions; however, there is no guarantee that the
details of _how_ AEL is compiled into extensions will remain stable.  In fact,
at the time of this commit, it has already changed twice, once in a
fundamental way.

Now normally, a new application would only be added to trunk.  However, this
application is explicitly to create a stable user-level API between versions,
and adding it to trunk only will not solve the user's problem of switching
between 1.6.2 and 1.8, nor will it help anybody switching from 1.8 to 1.10.
Therefore, it needs to go into existing release branches.  For the sake of
consistency, and also because one of the changes was between 1.4 and 1.6.x,
I am also electing to commit this to 1.4.

(closes issue #18910)
 Reported by: alexandrekeller
 Patches: 
       20110304__issue18919__1.6.2.diff.txt uploaded by tilghman (license 14)
       20110304__issue18919__1.4.diff.txt uploaded by tilghman (license 14)
 Tested by: alexandrekeller