"Caller*ID failed checksum" on Wildcard TDM2400P and TDM410
The last character in the caller id message is getting a framing error.

The checksum is the last character in the message.  A framing error in the
checksum could be because:
1) The sender did not send a full stop bit.
2) The sender cut off the FSK carrier too soon.
3) The sender opted to send zero of the specified zero to 10 trailing mark
bits and round-off errors in the code resulted in the code not being where
it thought it was in the demodulated bit stream.

Bit 8 of 'b' is set when parity error.
Bit 9 of 'b' is set when framing error.

Made ignore the framing and parity error bits if the errored character is
the checksum.  We can tolerate a framing/parity error there.  The checksum
character validates the message.

(closes issue #18474)
Reported by: nivek
Patches:
      callerid.c.1.patch uploaded by nivek (license 636) (with modifications)
Tested by: nivek