Prototype code creates a jitterbuffer on the read side of a channel using framehooks (awesomehooks)
This looks like it actually will work. Channels using a jitterbuffer will get read
called twice as often as those without, but it will just generate a NULL frame until
a real frame is time to be sent. In theory, this shouldn't be much more expensive
since NULL frames are just ignored.