CDR's are being written immediately on caller hangup.
CDR's are being written immediately on caller hangup.  The dialplan is not
able to modify it in the h exten.  The h exten in the initial context is
not run before closing CDR's when the bridge is unlinked if a macro is
active and does not have an h exten.

* Make ast_bridge_call() check for an h exten in the current context and
if a macro is active then the initial context.  The first h exten found is
then run before closing the CDR.

(closes issue #18212)
Reported by: leearcher
Patches:
      issue18212_v1.8.patch uploaded by rmudgett (license 664)
Tested by: rmudgett

Review: https://reviewboard.asterisk.org/r/1206/