This is a rewrite of the T.38 negotiation portion of the framehook code.
Things have been cleaned up a bit and code was added to handle as many
different T.38 negotiation scenarios as possible. Code was also added to
automatically disable the framehook when starting the gateway fails or when
T.38 negotiation fails. The gateway should now also automatically disable
itself if both channels support T.38 and T.38 negotiation is successful.

Now CED tones are only handled if the side receiving the CED tone indicates
that it supports T.38. A CED tone will trigger the gateway to send a T.38
reinvite to the party sending the fax.