When subscribing MWI to an unsolicited mailbox the first notification is incorrect.
A remote peer subscribed to MWI with the unsolicited option and a local
phone subscribed to the remote mailbox.  The notify message-summary events
are sent correctly except for the first one when subscribing, which will
always be 0.  This means the phone MWI indicator will be wrong until the
mailbox read/unread count changes and the event is fired.

Looks like this is a regression from ASTERISK-16149.

* Fix the logic to check the cache and if allowed then fallback to
manually counting mailbox messages.

(closes issue ASTERISK-17997)
Reported by: rsw686
Patches:
      jira_asterisk_17997_v1.8.patch (license #5621) uploaded by rmudgett
Tested by: rsw686

JIRA SWP-3551